Move whitelist application from business logic - i.e. the composition of the Notification - to email dispatch.
Notification business logic attempts to be agnostic with regard to the underlying dispatch and transport of notifications.  Initially the whitelist was applied on the business logic side, which means that it ought only operate on URIs.  This is inconvenient and opaque when a human is configuring a whitelist for User resources in Fedora.  The whitelist would have to reference Fedora URIs rather than simple email addresses or mailto URIs, making it difficult as a human to 1) compose a whitelist, and 2) interpret a whitelist and see which actual email address are whitelisted.

If whitelist processing were to remain on the business logic side, but be updated to operate on bare email addresses, that would be a nod to the underlying dispatch and transport of notifications as emails (i.e. the abstraction provided by the Dispatch API leaks).  Another option would have whitelist processing remain on the BL side, but be updated to resolve URIs to email addresses, and apply the whitelist.  But again, that breaks the abstraction if another form of notification transport were to be added (e.g. SMS or flashes in the view layer).

Therefore, this commit moves whitelist processing to the dispatch side of the house.  Instead of whitelist processing occurring in the Composer on BL side, it now occurs in the EmailDispatchImpl EmailComposer.  Whitelists are composed of bare email addresses and not opaque URIs.  If multiple dispatch transports are added later, the model (and format) of notification services configuration will need to be updated so each dispatch implementation can process its own form of whitelist, if recipient filtering is desired for these new implmentations.

There are some consequences of this move.  The logging aspect which operates at the Dispatch API join point is unaware of whitelist processing.  So when it logs the successful or failed delivery of a Notification, the recipients listed on the Notification may have been filtered inside the dispatch implementation.  This PR addresses this by logging the application of the whitelist at INFO level.  DEBUG level reveals more information about whitelist processing.

Finally, there is a dichotomy in the current BL/Dispatch impl split which has been there from the beginning, which is the "global cc" configuration field.  It is a bare email address, not a mailto address, and that should be addressed to be consistent with the whitelist processing (e.g. the "global cc" should either be updated to carry URIs, or processing of the "global cc" should be moved to the dispatch side of the house).

Addresses #55.